added loop assignment

+++ javaNotepad/assignments/assignmentMathsForm/HeightOfTriangle.java
@@ -2,10 +2,10 @@
 
 public class HeightOfTriangle {
     public static void main(String[] args) {
-        int base = 50;
-        int area = 500;
+        float base = 50;
+        float area = 500;
 
-        int height = 2*(area/base);
+        float height = 2*(area/base);
         System.out.println("The height of a triangle is: " +height);
     }
 }
